opti/requirements.txt
django==5.2.8
matplotlib==3.9.2
reportlab==4.2.5
Pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2


